Keyword to run action (see body)
There must be a specific set of unique words for the site to build and
deploy. It must be "build and deploy" wrapped in square brackets
("[]"). This is as a messure to reduce potential spam to Nekoweb's API.

This keyword should be included when merging a PR that updates anything
that will be on the site (static, content, images, headpages, views,
sometimes main.py depends on what it changes, etc).